+++ libcxx/src/memory.cpp
@@ -154,7 +154,7 @@ align(size_t alignment, size_t size, void*& ptr, size_t& space)
     {
         char* p1 = static_cast<char*>(ptr);
         char* p2 = (char*)((size_t)(p1 + (alignment - 1)) & -alignment);
-        ptrdiff_t d = p2 - p1;
+        size_t d = static_cast<size_t>(p2 - p1);
         if (d <= space - size)
         {
             r = p2;
